chinlite/pyproject.toml
[project]
name = "chinlite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.14",
]

[project.scripts]
chinlite = "chinlite:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
